Fix field types

+++ backend/src/main/java/com/batm/dto/TradeDTO.java
@@ -18,7 +18,7 @@ public class TradeDTO {
     private Integer type;
     private String paymentMethod;
     private BigDecimal margin;
-    private Integer minLimit;
-    private Integer maxLimit;
+    private Long minLimit;
+    private Long maxLimit;
     private String terms;
 }
